chat messages functional
