+++ src/styles/main.css
@@ -1,9 +1,25 @@
-#containerMessages, #chatLog {
+/* MESSAGES STYLING START */
+
+#containerMessages {
     border: 1px black dashed;
+    border-radius: 10px;
+    margin: 15px;
+    padding: 5px;
 }
 
 #chatLog {
+    padding-bottom: 10px;
     height: 100px;
     overflow-x: scroll;
     overflow-anchor: bottom;
 }
+
+#chatLogin {
+    display: block;
+    max-width: 100px;
+    margin: 50px auto 50px auto;
+    background-color: lightblue;
+    font-weight: bolder;
+}
+
+/* MESSAGES STYLING END */
